+++ includes/class-tmsm-frontend-optimizations.php
@@ -260,6 +260,9 @@ class Tmsm_Frontend_Optimizations {
 		// WooCommerce
 		$this->loader->add_action( 'woocommerce_scheduled_sales', $plugin_public, 'woocommerce_scheduled_sales_empty_cache', 200 );
 
+		// TAO Schedule Update
+		$this->loader->add_action( 'tao_publish_post', $plugin_public, 'tao_publish_post_emptycache', 200 );
+
 	}
 
 	/**